Major refactor of replay capability - CLI runs OK, one library test fails, all Android app tests fail
