+++ tec-library/bin/test/log4j.properties
@@ -1,0 +1,8 @@
+# The root logger with appender name 
+log4j.rootLogger = INFO, stdout
+  
+# Assign NAME a valid appender  
+log4j.appender.stdout = org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.target = System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout  
+log4j.appender.stdout.layout.conversionPattern=%20.20F(%4L): %m%n
